chore(sprvsr): moved types only used by k8supervisor to the appropriate file
