+++ sprvsr/k8s_supervisor.go
@@ -26,6 +26,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/negroni"
 
+	appsv1 "k8s.io/api/apps/v1"
 	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -57,6 +58,32 @@ type k8supervisor struct {
 
 	// A unique name (per k8s cluster) for your application to supervise
 	labelgenkey string
+}
+
+// ServiceNamer is called to determine what to name a Service Given a TrackedRepository
+type ServiceNamer func(repos.TrackedRepository) (string, error)
+
+// DeploymentNamer is called to determine what to name a Deployment Given a TrackedRepository
+type DeploymentNamer func(repos.TrackedRepository) (string, error)
+
+// DeploymentBuilder builds a Deployment based on the given TrackedRepository
+type DeploymentBuilder func(repos.TrackedRepository) (*appsv1.Deployment, error)
+
+// ServiceBuilder buidls a Service based on the given TrackedRepository
+type ServiceBuilder func(repos.TrackedRepository) (*apiv1.Service, error)
+
+// DeploymentPrep is called before building a deployment. This can be
+// used to provision additional resources before the Deployment is applied
+type DeploymentPrep func(repos.TrackedRepository) error
+
+// K8sConfiguration is a struct to describe the set of operations
+// a K8SSupervisor needs to manage a cluster
+type K8sConfiguration struct {
+	ServiceNamer      ServiceNamer
+	DeploymentNamer   DeploymentNamer
+	ServiceBuilder    ServiceBuilder
+	DeploymentBuilder DeploymentBuilder
+	PreDeploy         DeploymentPrep
 }
 
 // NewK8sSupervisor creates a new supervisor backed by Kubernetes
